Miscellaneous fixes

+++ Order.sql
@@ -2,13 +2,13 @@ CREATE TABLE order (
     PRIMARY KEY (order_id),
     FOREIGN KEY (agent_id)    REFERENCES agent(agent_id),
     FOREIGN KEY (customer_id) REFERENCES customer(customer_id),
-    order_id         INT          NOT NULL,
-    agent_id         INT          NOT NULL,
-    customer_id      INT          NOT NULL,
-    amount_due       INT          NOT NULL DEFAULT 0,
-    order_date       DATE         NOT NULL,
-    delivery_address VARCHAR(255) NOT NULL DEFAULT ' ',
-    gift             BOOLEAN      NOT NULL,
+    order_id         INT            NOT NULL,
+    agent_id         INT            NOT NULL,
+    customer_id      INT            NOT NULL,
+    amount_due       NUMERIC(19, 2) NOT NULL DEFAULT 0,
+    order_date       DATE           NOT NULL,
+    delivery_address VARCHAR(255)   NOT NULL DEFAULT ' ',
+    gift             BOOLEAN        NOT NULL,
 );
 
 CREATE TABLE ordered_product (
@@ -27,4 +27,3 @@ CREATE TABLE order_recipient (
     order_id  INT NOT NULL,
     person_id INT NOT NULL,
 );
-
